File corrupted
very frustrating. for some reason unknown to me the file stopped reading the csv. The only way I could get it to work is by rebuilding a new python file. Seems to be working again but I have spent the last 4 hours trying to figure this out. I will re-add the markdown later